pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
        "bcrypt>=5.0.0",
        "es-core-news-lg",
        "fastapi>=0.103.0",
        "google-generativeai==0.4.0",
        "gunicorn>=23.0.0",
        "numpy>=1.24.0,<2.0.0",
        "pandas>=2.3.3",
        "pdfplumber>=0.11.8",
        "psycopg2>=2.9.11",
        "pybloom-live>=4.0.0",
        "pydantic[email]>=2.0.0",
        "pyjwt>=2.10.1",
        "pymupdf>=1.26.6",
        "pypdf2==3.0.1",
        "python-dotenv==1.0.0",
        "python-multipart>=0.0.20",
        "qdrant-client==1.7.3",
        "spacy==3.7.4",
        "sqlalchemy==2.0.19",
        "torch>=2.9.0",
        "transformers>=4.57.1",
        "uvicorn[standard]==0.22.0",
]

[tool.uv.sources]
es-core-news-lg = { url = "https://github.com/explosion/spacy-models/releases/download/es_core_news_lg-3.7.0/es_core_news_lg-3.7.0.tar.gz" }
